requirements.txt
pbkdf2
toml
pycryptodome